DeskClock: fix crash during POWER_OFF_ALARM start up
When starting up as a resul of a POWER_OFF_ALARM, there is code in
the app to check if we just missed the alarm we are starting up for.
Since we start up two minutes early, this is usually not the case,
but there was a codepath while finishing the AlarmActivity that
assumed that mAlarmInstance wasn't null, so added a null check.

Also added some comments and renamed a method to make the purpose
of this startup path clearer to future generations.

CYNGNOS-1977

Change-Id: Icc67b71620b2859e596a46a04aff1e773637c694